chore(test): move test to new test suite Signed-off-by: Vladimir Lazar <vlazar@redhat.com>
